Bug 799701 - SX Editor error When Frequency=Once
When the frequency is once, in the SX edit frequency tab the scheduled
date is not marked. This is due to the 'first_date' variable not being
set and subsequent test for it being valid causing the calendar to be
cleared.

Add to the test for the recurrence type being 'once'.