Bug 799713 - Account Editor changes account color to gray when default selected.
Inadvertenly changed by 0a3fa2a because the logic is a bit too clever:
It's intentional that if the color-chooser color is the default that the
account color is set to NULL. That was removed for some flows and is
restored.

But that makes the default color button on the account edit dialog a
little misleading: It shows as gray but since selecting that makes the
account color NULL it's really transparent so we also enable alpha on the
color chooser and set the default color to be transparent so the
background checkerboard shows through.